refactor: KDA-28 Request 및 Response 패키지 구분
