+++ domain/src/main/java/com/upf464/koonsdiary/domain/di/CalendarUseCaseModule.kt
@@ -1,7 +1,7 @@
 package com.upf464.koonsdiary.domain.di
 
-import com.upf464.koonsdiary.domain.request.FetchCalendarRequest
-import com.upf464.koonsdiary.domain.response.FetchCalendarResponse
+import com.upf464.koonsdiary.domain.request.calendar.FetchCalendarRequest
+import com.upf464.koonsdiary.domain.response.calendar.FetchCalendarResponse
 import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import com.upf464.koonsdiary.domain.usecase.calendar.FetchCalendarUseCase
 import dagger.Binds
