+++ domain/src/main/java/com/upf464/koonsdiary/domain/di/DiaryUseCaseModule.kt
@@ -1,7 +1,11 @@
 package com.upf464.koonsdiary.domain.di
 
-import com.upf464.koonsdiary.domain.request.*
-import com.upf464.koonsdiary.domain.response.*
+import com.upf464.koonsdiary.domain.request.diary.*
+import com.upf464.koonsdiary.domain.response.EmptyResponse
+import com.upf464.koonsdiary.domain.response.diary.AddDiaryResponse
+import com.upf464.koonsdiary.domain.response.diary.AnalyzeSentimentResponse
+import com.upf464.koonsdiary.domain.response.diary.FetchDiaryResponse
+import com.upf464.koonsdiary.domain.response.diary.UpdateDiaryResponse
 import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import com.upf464.koonsdiary.domain.usecase.diary.*
 import dagger.Binds
