--- domain/src/main/java/com/upf464/koonsdiary/domain/request/AnalyzeSentimentRequest.kt
@@ -1,5 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-data class AnalyzeSentimentRequest(
-    val content: String
-) : Request
