--- domain/src/main/java/com/upf464/koonsdiary/domain/request/AutoSignInRequest.kt
@@ -1,3 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-object AutoSignInRequest : Request
