--- domain/src/main/java/com/upf464/koonsdiary/domain/request/DeleteDiaryRequest.kt
@@ -1,5 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-data class DeleteDiaryRequest(
-    val diaryId: Int
-) : Request
