--- domain/src/main/java/com/upf464/koonsdiary/domain/request/FetchCalendarRequest.kt
@@ -1,6 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-data class FetchCalendarRequest(
-    val year: Int,
-    val month: Int
-) : Request
