--- domain/src/main/java/com/upf464/koonsdiary/domain/request/FetchDiaryRequest.kt
@@ -1,5 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-data class FetchDiaryRequest(
-    val diaryId: Int
-) : Request
