--- domain/src/main/java/com/upf464/koonsdiary/domain/request/SignInWithKakaoRequest.kt
@@ -1,3 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-object SignInWithKakaoRequest : Request
