--- domain/src/main/java/com/upf464/koonsdiary/domain/request/SignInWithUsernameRequest.kt
@@ -1,6 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-data class SignInWithUsernameRequest(
-    val username: String,
-    val password: String
-) : Request
