--- domain/src/main/java/com/upf464/koonsdiary/domain/request/SignUpWithKakaoRequest.kt
@@ -1,3 +1,0 @@
-package com.upf464.koonsdiary.domain.request
-
-data class SignUpWithKakaoRequest(val nickname: String) : Request
