+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/calendar/FetchCalendarRequest.kt
@@ -1,0 +1,8 @@
+package com.upf464.koonsdiary.domain.request.calendar
+
+import com.upf464.koonsdiary.domain.request.Request
+
+data class FetchCalendarRequest(
+    val year: Int,
+    val month: Int
+) : Request
