+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/diary/AddDiaryRequest.kt
@@ -1,11 +1,11 @@
-package com.upf464.koonsdiary.domain.request
+package com.upf464.koonsdiary.domain.request.diary
 
 import com.upf464.koonsdiary.domain.model.Diary
 import com.upf464.koonsdiary.domain.model.Sentiment
+import com.upf464.koonsdiary.domain.request.Request
 import java.time.LocalDate
 
-data class UpdateDiaryRequest(
-    val diaryId: Int,
+data class AddDiaryRequest(
     val date: LocalDate,
     val content: String,
     val sentiment: Sentiment,
