+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/diary/AnalyzeSentimentRequest.kt
@@ -1,0 +1,7 @@
+package com.upf464.koonsdiary.domain.request.diary
+
+import com.upf464.koonsdiary.domain.request.Request
+
+data class AnalyzeSentimentRequest(
+    val content: String
+) : Request
