+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/diary/DeleteDiaryRequest.kt
@@ -1,0 +1,7 @@
+package com.upf464.koonsdiary.domain.request.diary
+
+import com.upf464.koonsdiary.domain.request.Request
+
+data class DeleteDiaryRequest(
+    val diaryId: Int
+) : Request
