+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/diary/FetchDiaryRequest.kt
@@ -1,0 +1,7 @@
+package com.upf464.koonsdiary.domain.request.diary
+
+import com.upf464.koonsdiary.domain.request.Request
+
+data class FetchDiaryRequest(
+    val diaryId: Int
+) : Request
