+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/user/SignInWithKakaoRequest.kt
@@ -1,0 +1,5 @@
+package com.upf464.koonsdiary.domain.request.user
+
+import com.upf464.koonsdiary.domain.request.Request
+
+object SignInWithKakaoRequest : Request
