+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/user/SignInWithUsernameRequest.kt
@@ -1,0 +1,8 @@
+package com.upf464.koonsdiary.domain.request.user
+
+import com.upf464.koonsdiary.domain.request.Request
+
+data class SignInWithUsernameRequest(
+    val username: String,
+    val password: String
+) : Request
