+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/user/SignUpWithKakaoRequest.kt
@@ -1,0 +1,5 @@
+package com.upf464.koonsdiary.domain.request.user
+
+import com.upf464.koonsdiary.domain.request.Request
+
+data class SignUpWithKakaoRequest(val nickname: String) : Request
