+++ domain/src/main/java/com/upf464/koonsdiary/domain/request/user/SignUpWithUsernameRequest.kt
@@ -1,4 +1,6 @@
-package com.upf464.koonsdiary.domain.request
+package com.upf464.koonsdiary.domain.request.user
+
+import com.upf464.koonsdiary.domain.request.Request
 
 data class SignUpWithUsernameRequest(
     val email: String,
