--- domain/src/main/java/com/upf464/koonsdiary/domain/response/AddDiaryResponse.kt
@@ -1,5 +1,0 @@
-package com.upf464.koonsdiary.domain.response
-
-data class AddDiaryResponse(
-    val diaryId: Int
-) : Response
