--- domain/src/main/java/com/upf464/koonsdiary/domain/response/UpdateDiaryResponse.kt
@@ -1,5 +1,0 @@
-package com.upf464.koonsdiary.domain.response
-
-data class UpdateDiaryResponse(
-    val diaryId: Int
-) : Response
