+++ domain/src/main/java/com/upf464/koonsdiary/domain/response/calendar/FetchCalendarResponse.kt
@@ -1,6 +1,7 @@
-package com.upf464.koonsdiary.domain.response
+package com.upf464.koonsdiary.domain.response.calendar
 
 import com.upf464.koonsdiary.domain.model.Sentiment
+import com.upf464.koonsdiary.domain.response.Response
 
 data class FetchCalendarResponse(
     val sentimentList: List<Sentiment?>
