+++ domain/src/main/java/com/upf464/koonsdiary/domain/response/diary/AddDiaryResponse.kt
@@ -1,0 +1,7 @@
+package com.upf464.koonsdiary.domain.response.diary
+
+import com.upf464.koonsdiary.domain.response.Response
+
+data class AddDiaryResponse(
+    val diaryId: Int
+) : Response
