+++ domain/src/main/java/com/upf464/koonsdiary/domain/response/diary/AnalyzeSentimentResponse.kt
@@ -1,6 +1,7 @@
-package com.upf464.koonsdiary.domain.response
+package com.upf464.koonsdiary.domain.response.diary
 
 import com.upf464.koonsdiary.domain.model.Sentiment
+import com.upf464.koonsdiary.domain.response.Response
 
 data class AnalyzeSentimentResponse(
     val sentiment: Sentiment
