+++ domain/src/main/java/com/upf464/koonsdiary/domain/response/diary/FetchDiaryResponse.kt
@@ -1,6 +1,7 @@
-package com.upf464.koonsdiary.domain.response
+package com.upf464.koonsdiary.domain.response.diary
 
 import com.upf464.koonsdiary.domain.model.Diary
+import com.upf464.koonsdiary.domain.response.Response
 
 data class FetchDiaryResponse(
     val diary: Diary
