+++ domain/src/main/java/com/upf464/koonsdiary/domain/response/diary/UpdateDiaryResponse.kt
@@ -1,0 +1,7 @@
+package com.upf464.koonsdiary.domain.response.diary
+
+import com.upf464.koonsdiary.domain.response.Response
+
+data class UpdateDiaryResponse(
+    val diaryId: Int
+) : Response
