+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/diary/AnalyzeSentimentUseCase.kt
@@ -3,8 +3,8 @@ package com.upf464.koonsdiary.domain.usecase.diary
 import com.upf464.koonsdiary.domain.common.DiaryValidator
 import com.upf464.koonsdiary.domain.error.DiaryError
 import com.upf464.koonsdiary.domain.repository.DiaryRepository
-import com.upf464.koonsdiary.domain.request.AnalyzeSentimentRequest
-import com.upf464.koonsdiary.domain.response.AnalyzeSentimentResponse
+import com.upf464.koonsdiary.domain.request.diary.AnalyzeSentimentRequest
+import com.upf464.koonsdiary.domain.response.diary.AnalyzeSentimentResponse
 import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import javax.inject.Inject
