+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/diary/FetchDiaryUseCase.kt
@@ -1,8 +1,8 @@
 package com.upf464.koonsdiary.domain.usecase.diary
 
 import com.upf464.koonsdiary.domain.repository.DiaryRepository
-import com.upf464.koonsdiary.domain.request.FetchDiaryRequest
-import com.upf464.koonsdiary.domain.response.FetchDiaryResponse
+import com.upf464.koonsdiary.domain.request.diary.FetchDiaryRequest
+import com.upf464.koonsdiary.domain.response.diary.FetchDiaryResponse
 import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import javax.inject.Inject
