+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/user/SignInWithKakaoUseCase.kt
@@ -4,7 +4,7 @@ import com.upf464.koonsdiary.domain.common.flatMap
 import com.upf464.koonsdiary.domain.common.handleWith
 import com.upf464.koonsdiary.domain.error.SignInError
 import com.upf464.koonsdiary.domain.repository.UserRepository
-import com.upf464.koonsdiary.domain.request.SignInWithKakaoRequest
+import com.upf464.koonsdiary.domain.request.user.SignInWithKakaoRequest
 import com.upf464.koonsdiary.domain.response.EmptyResponse
 import com.upf464.koonsdiary.domain.service.KakaoService
 import com.upf464.koonsdiary.domain.usecase.ResultUseCase
