+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/user/SignUpWithUsernameUseCase.kt
@@ -6,7 +6,7 @@ import com.upf464.koonsdiary.domain.common.flatMap
 import com.upf464.koonsdiary.domain.error.SignUpError
 import com.upf464.koonsdiary.domain.model.SignUpUser
 import com.upf464.koonsdiary.domain.repository.UserRepository
-import com.upf464.koonsdiary.domain.request.SignUpWithUsernameRequest
+import com.upf464.koonsdiary.domain.request.user.SignUpWithUsernameRequest
 import com.upf464.koonsdiary.domain.response.EmptyResponse
 import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import javax.inject.Inject
