+++ domain/src/test/java/com/upf464/koonsdiary/domain/usecase/user/AutoSignInUseCaseTest.kt
@@ -3,7 +3,7 @@ package com.upf464.koonsdiary.domain.usecase.user
 import com.upf464.koonsdiary.domain.error.SignInError
 import com.upf464.koonsdiary.domain.model.SignInType
 import com.upf464.koonsdiary.domain.repository.UserRepository
-import com.upf464.koonsdiary.domain.request.AutoSignInRequest
+import com.upf464.koonsdiary.domain.request.user.AutoSignInRequest
 import com.upf464.koonsdiary.domain.service.KakaoService
 import io.mockk.coEvery
 import io.mockk.coVerify
